+++ build.gradle
@@ -17,6 +17,7 @@ dependencies {
     implementation project(':module-api-config')
     implementation project(':module-domain-menu')
     implementation project(':module-domain-order')
+    implementation project(':module-api-menu')
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-jdbc'
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -59,6 +60,14 @@ project(':module-domain-menu') {
 project(':module-domain-order') {
     dependencies {
         implementation project(':common')
+        implementation project(':module-domain-menu')
+    }
+}
+
+project(':module-api-menu') {
+    dependencies {
+        implementation project(':common')
+        implementation project(':module-api-config')
         implementation project(':module-domain-menu')
     }
 }
