+++ module-api-menu/src/main/java/api/menu/dto/MenuGroupRequest.java
@@ -1,0 +1,16 @@
+package api.menu.dto;
+
+public class MenuGroupRequest {
+	private String name;
+
+	public MenuGroupRequest() {
+	}
+
+	public MenuGroupRequest(String name) {
+		this.name = name;
+	}
+
+	public String getName() {
+		return name;
+	}
+}
