+++ module-api-menu/src/main/java/api/menu/dto/MenuGroupResponse.java
@@ -1,0 +1,28 @@
+package api.menu.dto;
+
+import domain.menu.MenuGroup;
+
+public class MenuGroupResponse {
+	private Long id;
+	private String name;
+
+	public static MenuGroupResponse of(MenuGroup menuGroup) {
+		return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
+	}
+
+	public MenuGroupResponse() {
+	}
+
+	public MenuGroupResponse(Long id, String name) {
+		this.id = id;
+		this.name = name;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public String getName() {
+		return name;
+	}
+}
