+++ module-api-menu/src/main/java/api/menu/ui/MenuControllerAdvice.java
@@ -1,0 +1,8 @@
+package api.menu.ui;
+
+import api.config.BaseControllerAdvice;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+
+@ControllerAdvice(basePackages = "api.menu.ui")
+public class MenuControllerAdvice extends BaseControllerAdvice {
+}
