+++ module-api-menu/src/main/java/api/menu/ui/MenuGroupRestController.java
@@ -1,0 +1,34 @@
+package api.menu.ui;
+
+import api.menu.application.MenuGroupService;
+import api.menu.dto.MenuGroupRequest;
+import api.menu.dto.MenuGroupResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class MenuGroupRestController {
+	private final MenuGroupService menuGroupService;
+
+	public MenuGroupRestController(final MenuGroupService menuGroupService) {
+		this.menuGroupService = menuGroupService;
+	}
+
+	@PostMapping("/api/menu-groups")
+	public ResponseEntity<MenuGroupResponse> create(@RequestBody final MenuGroupRequest request) {
+		final MenuGroupResponse created = menuGroupService.create(request);
+		final URI uri = URI.create("/api/menu-groups/" + created.getId());
+		return ResponseEntity.created(uri).body(created);
+	}
+
+	@GetMapping("/api/menu-groups")
+	public ResponseEntity<List<MenuGroupResponse>> list() {
+		return ResponseEntity.ok().body(menuGroupService.list());
+	}
+}
