+++ module-api-menu/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+logging.level.org.springframework.jdbc.core=TRACE
+spring.h2.console.enabled=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=create
+spring.flyway.enabled=false
