+++ module-api-menu/src/test/java/api/menu/ApplicationTest.java
@@ -1,0 +1,11 @@
+package api.menu;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTest {
+	@Test
+	void contextLoads() {
+	}
+}
