+++ settings.gradle
@@ -3,4 +3,5 @@ include 'common'
 include 'module-api-config'
 include 'module-domain-menu'
 include 'module-domain-order'
+include 'module-api-menu'
