--- src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,16 +1,0 @@
-package kitchenpos.menu.dto;
-
-public class MenuGroupRequest {
-	private String name;
-
-	public MenuGroupRequest() {
-	}
-
-	public MenuGroupRequest(String name) {
-		this.name = name;
-	}
-
-	public String getName() {
-		return name;
-	}
-}
