--- src/main/java/kitchenpos/menu/dto/MenuGroupResponse.java
@@ -1,28 +1,0 @@
-package kitchenpos.menu.dto;
-
-import domain.menu.MenuGroup;
-
-public class MenuGroupResponse {
-	private Long id;
-	private String name;
-
-	public static MenuGroupResponse of(MenuGroup menuGroup) {
-		return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
-	}
-
-	public MenuGroupResponse() {
-	}
-
-	public MenuGroupResponse(Long id, String name) {
-		this.id = id;
-		this.name = name;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-}
