--- src/main/java/kitchenpos/menu/ui/MenuControllerAdvice.java
@@ -1,8 +1,0 @@
-package kitchenpos.menu.ui;
-
-import api.config.BaseControllerAdvice;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice(basePackages = "kitchenpos.menu.ui")
-public class MenuControllerAdvice extends BaseControllerAdvice {
-}
