--- src/main/java/kitchenpos/menu/ui/ProductRestController.java
@@ -1,34 +1,0 @@
-package kitchenpos.menu.ui;
-
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.dto.ProductRequest;
-import kitchenpos.menu.dto.ProductResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-public class ProductRestController {
-	private final ProductService productService;
-
-	public ProductRestController(final ProductService productService) {
-		this.productService = productService;
-	}
-
-	@PostMapping("/api/products")
-	public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest productRequest) {
-		final ProductResponse created = productService.create(productRequest);
-		final URI uri = URI.create("/api/products/" + created.getId());
-		return ResponseEntity.created(uri).body(created);
-	}
-
-	@GetMapping("/api/products")
-	public ResponseEntity<List<ProductResponse>> list() {
-		return ResponseEntity.ok().body(productService.list());
-	}
-}
